copperplate: Address compiler warning in pvheapobj_validate
Resolves

include/copperplate/heapobj.h:180:22: warning: conversion to ‘size_t {aka long unsigned int}’ from ‘ssize_t {aka long int}’ may change the sign of the result [-Wsign-conversion]

+++ include/copperplate/heapobj.h
@@ -177,7 +177,7 @@ static inline
 size_t pvheapobj_validate(struct heapobj *hobj, void *ptr)
 {
 	ssize_t size = heapmem_check((struct heap_memory *)hobj->pool, ptr);
-	return size < 0 ? 0 : size;
+	return size < 0 ? 0 : (size_t)size;
 }
 
 static inline
